feat: refactor prisma imports and update tsconfig include paths; add prisma.js for client initialization
