+++ apps/api/src/prisma/prisma.js
@@ -1,0 +1,5 @@
+'use strict';
+Object.defineProperty(exports, '__esModule', { value: true });
+const prisma_1 = require('prisma-client-7a5d7cf047f46244bd13fa0ec2311c64db7532942d10ac286efb5d925a2a8be0');
+const prisma = new prisma_1.PrismaClient();
+exports.default = prisma;
